#53 refactor: Message 타입 createdAt 필드 제거 및 관련 파일명 변경
