+++ src/types/Chat.ts
@@ -5,5 +5,4 @@ export type Message = {
   role: "user" | "bot";
   content: string;
   type?: MessageType; // 기본 "text"
-  createdAt?: string; // 메시지 시간
 };
